Revise my testing framework to use HtmlUnit (not HttpUnit!) underneath, instead of my own ReadTag program, for getting HTML tags.
Because of this there are new versions of A LOT of Jar files, so this needs
to be tested (of course, this is a test suite upgrade). Please update soon and
let me know of any anomalies!

This also continues the plan of dividing WEB-INF/lib from ./lib; the former for things that get deployed, the latter for things that do not, or are not needed in
the deployed package.